separate CreateAttempt from TaskDetailsToolbar
